# packages/api/dist/swagger/output.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.1.0",
        "title": "Todo List API",
        "description": "Api to retrieve, create and toggle todo list items"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/todo-list-items": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "text": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/todo-list-items/{id}/toggle": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "ToDoListItems": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "Plan visit to friends"
                    },
                    "done": {
                        "type": "boolean",
                        "example": false
                    }
                }
            }
        }
    }
}